+++ R/compareFigures.R
@@ -10,12 +10,14 @@
 #' @param .path_previous file or directory path to compare .path_base to (leave as NULL to compare to repo version)
 #' @param .file_types file extensions to include in comparison (only pdf and png allowed)
 #' @param .side_by_side Logical. Should outputs be displayed side by side?
+#' @param .show_on_load Logical. Should all figures be open when the page loads? (Set to FALSE for slow loading pages with many figures.)
 #' 
 #' @export
 compareFigures <- function(.path_current, 
                            .path_previous = NULL,
                            .file_types = "pdf",
-                           .side_by_side = TRUE) {
+                           .side_by_side = TRUE,
+                           .show_on_load = TRUE) {
   
   .allowed_exts <- c("png", "pdf")
   
@@ -29,14 +31,16 @@ compareFigures <- function(.path_current,
     compareModified(
       .path = .path_current, 
       .side_by_side = .side_by_side, 
-      .file_exts = .file_types
+      .file_exts = .file_types,
+      .show_on_load = .show_on_load
     )
   } else {
     compareLocal(
       .path_base = .path_previous, 
       .path_compare = .path_current, 
       .side_by_side = .side_by_side,
-      .file_exts = .file_types
+      .file_exts = .file_types,
+      .show_on_load = .show_on_load
     )
   }