+++ R/compareLocal.R
@@ -8,8 +8,10 @@
 #' 
 #' @param .file_exts file extensions to include in comparison
 #' 
+#' @param .show_on_load Logical. Should all figures be open when the page loads?
+#' 
 #' @noRd
-compareLocal <- function(.path_base, .path_compare, .side_by_side = TRUE, .file_exts = c("png", "pdf", "tex")) {
+compareLocal <- function(.path_base, .path_compare, .side_by_side = TRUE, .file_exts = c("png", "pdf", "tex"), .show_on_load = TRUE) {
    
   .allowed_exts <- c("png", "pdf", "tex")
   
@@ -75,6 +77,7 @@ compareLocal <- function(.path_base, .path_compare, .side_by_side = TRUE, .file_
   buildCompare(
     .dfpaths = .dfpaths, 
     .side_by_side = .side_by_side,
-    .headings = c(.path_base, .path_compare)
+    .headings = c(.path_base, .path_compare),
+    .show_on_load = .show_on_load
   )
 }